python_language/requirements.txt
annotated-types==0.6.0
arrow==1.3.0
attrs==23.1.0
colorama==0.4.6
colour==0.1.5
contourpy==1.2.0
cycler==0.12.1
fonttools==4.45.0
greenlet==3.0.1
kiwisolver==1.4.5
marshmallow==3.20.1
matplotlib==3.8.2
mimesis==11.1.0
mpmath==1.3.0
networkx==3.2.1
npyscreen==4.10.5
number-parser==0.3.2
numpy==1.26.1
packaging==23.2
Pillow==10.2.0
prettytable==3.9.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
PyQt6==6.5.3
PyQt6-Qt6==6.5.3
PyQt6-sip==13.6.0
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.23
sqlitedict==2.1.0
sympy==1.12
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
wcwidth==0.2.12
faker==21.0.0
russian-names
environs
scipy
automata-lib
coloraide
pandas

